requirements.txt
# FastAPI and Server
fastapi==0.119.0
uvicorn[standard]==0.37.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
geoalchemy2==0.15.0

# Data Processing
pandas==2.2.3
numpy==2.1.3

# GIS and Spatial
shapely==2.0.5
geopandas==0.15.2
geopy==2.4.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
slowapi==0.1.9

# HTTP Client
httpx==0.27.2
aiohttp==3.11.3

# Validation
pydantic==2.9.2
pydantic-settings==2.6.1

# Utilities
python-dateutil==2.9.0.post0

# Optional: Redis for distributed state (optional, falls back to local state)
redis==5.0.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
pytest-mock==3.14.0
httpx==0.27.2
faker==28.0.0
locust==2.24.1  # Load testing

# Security & Validation
bleach==6.1.0
cryptography==43.0.1

# Structured Logging
structlog==24.2.0
python-json-logger==2.0.7

# Monitoring
prometheus-client==0.20.0

# Note: For production GIS data integration, consider adding:
# geopandas (already included)
# fiona (for reading shapefiles)
# rtree (for spatial indexing)
# For Natural Earth data: Download from naturalearthdata.com
